Radio and SPI is behaving now
When the radio was in sleep mode it may not have been properly
working. This test writes values to the AES key registers and then
attempts to read the values back out. I was successful with this
test. The documentation doesn't really state if there are any
functional differences between sleep mode and standby mode. The
most I can tell is a difference in time it takes to go to other
modes.